feat(procedures): implemented gdpr form
ref:MANAGER-15317
