+++ packages/manager/apps/procedures/src/components/PageLayout/PageLayout.component.tsx
@@ -11,7 +11,7 @@ export const PageLayout: FunctionComponent<Props> = ({ children }) => (
       <div className="inline-block pb-6 md:pb-12">
         <img src={ovhCloudLogo} alt="ovh-cloud-logo" className="app-logo" />
       </div>
-      <div className="flex justify-center app-content lg:w-8/12 mx-auto min-h-[500px] sm:shadow sm:border-none border-t-[1px] border-gray-300 px-6">
+      <div className="flex justify-center app-content lg:w-8/12 mx-auto min-h-[500px] sm:shadow-[0_0_6px_0_rgba(40,89,192,0.2)] sm:border-none border-t-[1px] border-gray-300 px-6">
         <div className="md:p-8 w-full">{children}</div>
       </div>
     </div>
