+++ packages/manager/apps/procedures/src/pages/rgdp/rgdpForm/SelectField/SelectField.component.tsx
@@ -1,7 +1,21 @@
+import {
+  OsdsFormField,
+  OsdsSelect,
+  OsdsSelectOption,
+  OsdsText,
+} from '@ovhcloud/ods-components/react';
+import {
+  ODS_THEME_COLOR_INTENT,
+  ODS_THEME_TYPOGRAPHY_LEVEL,
+} from '@ovhcloud/ods-common-theming';
 import { Control, Controller, Validate, ValidationRule } from 'react-hook-form';
 import React, { FunctionComponent } from 'react';
-import { Select, SelectOption } from '@/components/Select/Select.component';
 import { GDPRFormValues } from '@/types/gdpr.type';
+
+export type SelectOption = {
+  label: string;
+  value: string;
+};
 
 type Props = {
   name: keyof GDPRFormValues;
@@ -27,6 +41,8 @@ export const SelectField: FunctionComponent<Props> = ({
   options,
   placeholder,
 }) => {
+  const id = `field_id_${name}`;
+
   return (
     <div>
       <Controller
@@ -41,18 +57,35 @@ export const SelectField: FunctionComponent<Props> = ({
           field: { onChange, onBlur, value, name: _name },
           fieldState: { error },
         }) => (
-          <Select
-            label={label}
-            name={_name}
-            required={Boolean(required)}
-            id={`field_id_${_name}`}
-            onBlur={onBlur}
-            onChange={onChange}
-            value={value}
-            options={options}
-            placeholder={placeholder}
-            error={error?.message}
-          />
+          <OsdsFormField error={error?.message}>
+            {label && (
+              <label htmlFor={id} slot="label">
+                <OsdsText
+                  level={ODS_THEME_TYPOGRAPHY_LEVEL.heading}
+                  color={ODS_THEME_COLOR_INTENT.primary}
+                >
+                  {label}:
+                </OsdsText>
+              </label>
+            )}
+            <OsdsSelect
+              onOdsValueChange={onChange}
+              onBlur={onBlur}
+              value={value}
+              name={name}
+              required={Boolean(required)}
+              id={id}
+              error={Boolean(error) || undefined}
+              inline
+            >
+              {placeholder && <span slot="placeholder">{placeholder}</span>}
+              {options.map((option, index) => (
+                <OsdsSelectOption value={option.value} key={index}>
+                  {option.label}
+                </OsdsSelectOption>
+              ))}
+            </OsdsSelect>
+          </OsdsFormField>
         )}
       />
     </div>
